AUTO-102: Move configuration to own package
- We are getting quite a few configuration classes now so create a new
  package for them
