+++ hub/policy/src/main/java/uk/gov/ida/hub/policy/PolicyModule.java
@@ -13,6 +13,8 @@ import uk.gov.ida.eventsink.EventSinkProxy;
 import uk.gov.ida.hub.policy.annotations.Config;
 import uk.gov.ida.hub.policy.annotations.SamlEngine;
 import uk.gov.ida.hub.policy.annotations.SamlSoapProxy;
+import uk.gov.ida.hub.policy.configuration.AssertionLifetimeConfiguration;
+import uk.gov.ida.hub.policy.configuration.PolicyConfiguration;
 import uk.gov.ida.hub.policy.controllogic.AuthnRequestFromTransactionHandler;
 import uk.gov.ida.hub.policy.controllogic.ResponseFromIdpHandler;
 import uk.gov.ida.hub.policy.domain.AssertionRestrictionsFactory;
