+++ hub/policy/src/main/java/uk/gov/ida/hub/policy/domain/controller/AbstractMatchRequestSentStateController.java
@@ -3,7 +3,7 @@ package uk.gov.ida.hub.policy.domain.controller;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.ida.hub.policy.PolicyConfiguration;
+import uk.gov.ida.hub.policy.configuration.PolicyConfiguration;
 import uk.gov.ida.hub.policy.contracts.AbstractAttributeQueryRequestDto;
 import uk.gov.ida.hub.policy.domain.MatchFromMatchingService;
 import uk.gov.ida.hub.policy.domain.NoMatchFromMatchingService;
