+++ hub/policy/src/main/java/uk/gov/ida/hub/policy/domain/controller/EidasAwaitingCycle3DataStateController.java
@@ -2,7 +2,7 @@ package uk.gov.ida.hub.policy.domain.controller;
 
 import com.google.common.base.Optional;
 import org.joda.time.DateTime;
-import uk.gov.ida.hub.policy.PolicyConfiguration;
+import uk.gov.ida.hub.policy.configuration.PolicyConfiguration;
 import uk.gov.ida.hub.policy.contracts.EidasAttributeQueryRequestDto;
 import uk.gov.ida.hub.policy.contracts.MatchingServiceConfigEntityDataDto;
 import uk.gov.ida.hub.policy.domain.AssertionRestrictionsFactory;
