+++ hub/policy/src/main/java/uk/gov/ida/hub/policy/domain/controller/MatchRequestSentStateController.java
@@ -1,7 +1,7 @@
 package uk.gov.ida.hub.policy.domain.controller;
 
 import org.joda.time.DateTime;
-import uk.gov.ida.hub.policy.PolicyConfiguration;
+import uk.gov.ida.hub.policy.configuration.PolicyConfiguration;
 import uk.gov.ida.hub.policy.contracts.AttributeQueryRequestDto;
 import uk.gov.ida.hub.policy.contracts.MatchingServiceConfigEntityDataDto;
 import uk.gov.ida.hub.policy.domain.PersistentId;
