+++ hub/policy/src/test/java/uk/gov/ida/hub/policy/domain/controller/EidasUserAccountCreationRequestSentStateControllerTest.java
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.ida.hub.policy.PolicyConfiguration;
+import uk.gov.ida.hub.policy.configuration.PolicyConfiguration;
 import uk.gov.ida.hub.policy.domain.ResponseFromHubFactory;
 import uk.gov.ida.hub.policy.domain.ResponseFromMatchingService;
 import uk.gov.ida.hub.policy.domain.State;
